Renamed AggregationQueue to Multiplexer and removed old Multiplexer functionalities. Now, scheduler will do most things
