+++ src/coreL2/Multiplexer/Multiplexer.cpp
@@ -7,7 +7,7 @@
 @Arquive name : Multiplexer.cpp
 @Classification : Multiplexer
 @
-@Last alteration : February 13th, 2020
+@Last alteration : March 11th, 2020
 @Responsible : Eduardo Melao
 @Email : emelao@cpqd.com.br
 @Telephone extension : 7015
@@ -19,142 +19,162 @@ Company : Centro de Pesquisa e Desenvolvimento em Telecomunicacoes (CPQD)
 Direction : Diretoria de Operações (DO)
 UA : 1230 - Centro de Competencia - Sistemas Embarcados
 
-@Description : This module manages the queues of SDUs that will generate PDUs 
-    to send to destination.
+@Description : This module manages a single queue of aggregatedSDUs and 
+    is used on encoding and decoding of the MAC PDU. 
 */
 
 #include "Multiplexer.h"
 
 Multiplexer::Multiplexer(
-    int _numberDestinations,        //Number of destinations with PDUs to aggregate
-    uint8_t _sourceMac,             //Source MAC Address
-    uint8_t* _destinationMacX,      //Destination MAC Addresses
-    uint16_t* _maxNumberBytesX,     //Maximum number of Bytes for each Buffer of PDUs
+    int _maxNumberBytes,            //Maximum number of Bytes in a PDU
+    uint8_t _sourceAddress,         //Source MAC Address
+    uint8_t _destinationAddress,    //Destination MAC Address
     bool _verbose)                  //Verbosity flag
 {
-    numberDestinations = _numberDestinations;
-    sourceMac = _sourceMac;
-    destinationMacX = _destinationMacX;
-    maxNumberBytesX = _maxNumberBytesX;
+    //Initialize class variables
+    maxNumberBytes = _maxNumberBytes;
+    buffer = new char[maxNumberBytes];
+    sourceAddress = _sourceAddress;
+    destinationAddress = _destinationAddress;
+    numberSDUs = 0;
+    controlOffset = 0;
     verbose = _verbose;
-
-    //Alloc array of number of Bytes aggregated and AggregationQueues for each destination
-    numberBytesAgrX = new uint16_t[numberDestinations];
-    aggregationQueueX = new AggregationQueue*[numberDestinations];
-
-    //Initialize each position of arrays
-    for(int i=0;i<numberDestinations;i++){
-    	numberBytesAgrX[i] = 0;
-        aggregationQueueX[i] = new AggregationQueue(maxNumberBytesX[i], sourceMac, destinationMacX[i], verbose);
-    }
-    if(_verbose) cout<<"[Multiplexer] Created successfully."<<endl;
 }
 
-Multiplexer::~Multiplexer()
+Multiplexer::Multiplexer(
+    char* _buffer,                      //Buffer containing PDU for decoding
+    uint8_t _numberSDUs,                //Number of SDUs into PDU
+    uint16_t* _sizesSDUs,               //Array of sizes of each SDU
+    uint8_t* _flagsDataControlSDUS,     //Array of Data/Control flags
+    bool _verbose)                      //Verbosity flag
 {
-    for(int i=0;i<numberDestinations;i++)
-        delete aggregationQueueX[i];
-    delete[] aggregationQueueX;
-    delete[] numberBytesAgrX;
+    offset = 0;
+    buffer = _buffer;
+    numberSDUs = _numberSDUs;
+    verbose = _verbose; 
+
+    //Copy Size and Flag Data/Control arrays
+    for(int i=0;i<numberSDUs;i++){
+        sizesSDUs.push_back(_sizesSDUs[i]);
+        flagsDataControlSDUs.push_back(_flagsDataControlSDUS[i]);
+    }
+
+    //Delete previously dynamicaly allocated arrays
+    delete [] _sizesSDUs;
+    delete [] _flagsDataControlSDUS;
 }
 
-void 
-Multiplexer::addSdu(
-    char* sdu,                  //SDU buffer
-    uint16_t size,              //Number of Bytes of SDU
-    uint8_t flagDataControl,    //Data/Control flag
-    uint8_t destinationMac)     //Destination MAC Address
-{
-    int index = getAggregationQueueIndex(destinationMac);
-
-    //Test if index is valid
-    if(index==-1){
-        if(verbose) cout<<"[Multiplexer] Bad MAC Address found trying to Add SDU to Aggregation Queue."<<endl;
-        return;
-    }
-
-    //Test if queue is full
-    if(aggregationQueueFull(destinationMac, size)){
-        if(verbose) cout<<"[Multiplexer] Number of bytes exceed buffer max length."<<endl;
-        return;
-    }
-
-    //Add SDU to AggregationQueue
-    if(aggregationQueueX[index]->addSDU(sdu, size, flagDataControl)){
-        numberBytesAgrX[index]+=size;
-        if(verbose&&flagDataControl) cout<< "[Multiplexer] Data SDU added to queue!"<<endl;
-        if(verbose&&!flagDataControl) cout<< "[Multiplexer] Control SDU added to queue!"<<endl;
-    }
-}
-
-ssize_t 
-Multiplexer::getPdu(
-    char* buffer,       //Buffer to store PDU
-    uint8_t macAddress) //Destination MAC Address of PDU
-{
-    ssize_t size;   //Size of PDU
-    int index = getAggregationQueueIndex(macAddress);
-
-    //Test if index is valid
-    if(index==-1){
-        if(verbose) cout<<"[Multiplexer] Bad MAC Address found trying to Get SDU from Aggregation Queue."<<endl;
-        return -1;
-    }
-
-    //Test if there are bytes to return
-    if(numberBytesAgrX[index] == 0){
-        if(verbose) cout<<"[Multiplexer] Could not get PDU: no Bytes to transfer."<<endl;
-        return -1;
-    }
-
-    //Creates a ProtocolPackage to receive the PDU
-    ProtocolPackage* pdu = aggregationQueueX[index]->getPDUPackage();
-
-    if(verbose) cout<<"[Multiplexer] Inserting MAC Header."<<endl;
-
-    //Inserts MacHeader and returns PDU size
-    pdu->insertMacHeader();
-    size = pdu->getPduSize();
-    memcpy(buffer, pdu->buffer, size);
-    delete pdu;
-
-    return size;
+Multiplexer::~Multiplexer(){
+    delete[] buffer;
 }
 
 int 
-Multiplexer::getNumberDestinations(){
-    return numberDestinations;
+Multiplexer::currentBufferLength(){
+    int length = 0;
+    for(int i=0;i<numberSDUs;i++){
+        length += sizesSDUs[i];
+    }
+    return length;
 }
 
-bool
-Multiplexer::aggregationQueueFull(
-    uint8_t macAddress,     //Destination MAC Address
-    uint16_t sduSize)       //Size of SDU to be enqueued
-{
-    int index = getAggregationQueueIndex(macAddress);
+int Multiplexer::getNumberofBytes(){
+    int numberBytes = 0;
+    //Header length:
+    numberBytes = 2 + 2*numberSDUs;
 
-    //Verify bad macAddress
-    if(index==-1){
-        if(verbose) cout<<"[Multiplexer] MAC Address not found verifying if Aggregation Queue is full"<<endl;
-        exit(5); 
-    }
-
-    //Returns if size of:
-    //  -> sduSize: Size of SDU that wants to be added by Multiplexer
-    //  -> 2: CRC
-    //  -> 2: Header additions - Size (15bits) + Flag D/C (1 bit)
-    //  -> AggregationQueue::getNumberofBytes(): Size of SDUs aggregated + Header of them
-    //is bigger tham maximum number of bytes
-    return ((sduSize + 2 + 2 + aggregationQueueX[index]->getNumberofBytes())>maxNumberBytesX[index]);
+    //Buffer length:
+    numberBytes +=currentBufferLength();
+    return numberBytes;
 }
 
-int
-Multiplexer::getAggregationQueueIndex(
-    uint8_t macAddress)     //Destination MAC Address
+bool 
+Multiplexer::addSduPosition(
+    char* sdu,                  //Buffer containing single SDU
+    uint16_t size,              //SDU size
+    uint8_t flagDataControl,    //SDU Data/Control flag
+    int position)               //Position in the queue where SDU will be added
 {
-    for(int i=0;i<numberDestinations;i++){
-        if(destinationMacX[i]==macAddress)
-            return i;
+    int bufferOffset = 0;                   //Buffer offset to manage copying arrays
+    int length = currentBufferLength();     //Current length of the queue
+
+    //Resize vectors:
+    sizesSDUs.resize(numberSDUs+1);
+    flagsDataControlSDUs.resize(numberSDUs+1);
+
+    //Copying forward information already in queue
+    for(int i=numberSDUs-1;i>=position;i--){
+        sizesSDUs[i+1] = sizesSDUs[i];
+        flagsDataControlSDUs[i+1] = flagsDataControlSDUs[i];
+        bufferOffset+=sizesSDUs[i+1];
     }
-    return -1;
+    for(int i=(length-1);i>=(length-bufferOffset);i--){
+        buffer[i+size] = buffer[i];
+    }
+
+    //Implanting actual SDU
+    sizesSDUs[position] = size;
+    flagsDataControlSDUs[position] = flagDataControl;
+    for(int i=0;i<size;i++)
+        buffer[length-bufferOffset+i] = sdu[i];
+    numberSDUs++;
+    if(!flagDataControl) controlOffset++;
+    if(verbose) cout<<"[Multiplexer] Multiplexed "<<(int)numberSDUs<<" SDUs into PDU."<<endl;
+    return true;
+}
+
+bool 
+Multiplexer::addSDU(
+    char* sdu,                  //Buffer containing single SDU
+    uint16_t size,              //SDU size
+    uint8_t flagDataControl)    //SDU Data/Control flag
+{
+    //Verify if it is possible to insert SDU (considering CRC)
+    if((size+2+getNumberofBytes())>maxNumberBytes){
+        if(verbose) cout<<"[Multiplexer] Tried to multiplex SDU which size extrapolates maxNumberBytes."<<endl;
+        exit(4);
+    }
+
+    //Adds SDU to position depending if it is Data or Control SDU
+    return flagDataControl? addSduPosition(sdu, size, flagDataControl, numberSDUs):addSduPosition(sdu, size, flagDataControl, controlOffset);
+}
+
+ssize_t 
+Multiplexer::getSDU(
+    char* sdu)      //Buffer to store SDU
+{
+    //Test if Decoding queue has ended
+    if(offset == numberSDUs){
+        if(verbose) cout<<"[Multiplexer] End of demultiplexing."<<endl;
+        return -1;
+    }
+
+    //Set an position offset to next SDU position in buffer
+    int positionBuffer = 0;
+    for(int i=0;i<offset;i++)
+        positionBuffer+=sizesSDUs[i];
+    
+    //Copy SDU from buffer
+    for(int i=0;i<sizesSDUs[offset];i++)
+        sdu[i] = buffer[positionBuffer+i];
+    
+    //Increment decoding offset
+    offset++;
+    if(verbose) cout<<"[Multiplexer] Demultiplexed SDU "<<offset<<endl;
+    return sizesSDUs[offset-1];
+}
+
+ProtocolPackage* 
+Multiplexer::getPDUPackage(){
+    ProtocolPackage* pdu = new ProtocolPackage(sourceAddress, destinationAddress, numberSDUs, &(sizesSDUs[0]), &(flagsDataControlSDUs[0]), buffer, verbose);
+    return pdu;
+}
+
+uint8_t 
+Multiplexer::getCurrentDataControlFlag(){
+    return flagsDataControlSDUs[offset-1];
+}
+
+uint8_t 
+Multiplexer::getDestinationAddress(){
+    return destinationAddress;
 }
